Moved local_path function into settings.py
