+++ metadata-mapper/metadata_mapper/mapper.py
@@ -14,16 +14,6 @@ class VernacularReader(object):
         self.collection_id = payload.get('collection_id')
         self.page_filename = payload.get('page_filename')
 
-    def local_path(self, folder):
-        parent_dir = os.sep.join(os.getcwd().split(os.sep)[:-1])
-        local_path = os.sep.join([
-            parent_dir,
-            'rikolti_bucket',
-            folder,
-            str(self.collection_id),
-        ])
-        return local_path
-
     def get_api_response(self):
         if settings.DATA_SRC == 'local':
             return self.get_local_api_response()
@@ -31,7 +21,8 @@ class VernacularReader(object):
             return self.get_s3_api_response()
 
     def get_local_api_response(self):
-        local_path = self.local_path('vernacular_metadata')
+        local_path = settings.local_path(
+            'vernacular_metadata', self.collection_id)
         page_path = os.sep.join([local_path, str(self.page_filename)])
         page = open(page_path, "r")
         api_response = page.read()
@@ -51,18 +42,9 @@ class UCLDCWriter(object):
         self.collection_id = payload.get('collection_id')
         self.page_filename = payload.get('page_filename')
 
-    def local_path(self, folder):
-        parent_dir = os.sep.join(os.getcwd().split(os.sep)[:-1])
-        local_path = os.sep.join([
-            parent_dir,
-            'rikolti_bucket',
-            folder,
-            str(self.collection_id),
-        ])
-        return local_path
-
     def write_local_mapped_metadata(self, mapped_metadata):
-        local_path = self.local_path('mapped_metadata')
+        local_path = settings.local_path(
+            'mapped_metadata', self.collection_id)
         if not os.path.exists(local_path):
             os.makedirs(local_path)
         page_path = os.sep.join([local_path, str(self.page_filename)])
